+++ plugins/core/qRANSAC_SD/RANSAC_SD_orig/Sphere.h
@@ -107,7 +107,7 @@ private:
 		void Derivatives(const ScalarType *params, IteratorT begin, IteratorT end,
 			const ScalarType *values, const ScalarType *temp, ScalarType *matrix) const
 		{
-			int size = end - begin;
+			int size = static_cast<int>(end - begin);
 #ifdef DOPARALLEL
 			#pragma omp parallel for schedule(static)
 #endif
